backend/requirements.txt
flask
flask-cors
flask-mail
flask-mysqldb
mysqlclient
pandas
apscheduler
numpy
openpyxl
PyJWT
arrow
pytz
Werkzeug
groq
python-dotenv